refactor: 보드에서 기물을 초기화하는 기능 이동
